+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
-v4.16.0 (2022-10-??)
+v4.16.0 (2022-10-28)
 --------------------
 
 Added